Revert "Add more stubs"
This reverts commit 6947256bf9e2e8ffef774793d5de88df504c3e10.
