+++ src/CSHTML5.Runtime/CSHTML5.Runtime.csproj
@@ -52,21 +52,12 @@
   <ItemGroup>
     <Compile Include="Core\Main\INTERNAL_DefaultValueStore.cs" />
     <Compile Include="Core\Other\INTERNAL_DispatcherQueueHandler.cs" />
-    <Compile Include="System.Collections.ObjectModel\WORKINPROGRESS\ReadOnlyObservableCollection_1.cs" />
     <Compile Include="System.ComponentModel\IPagedCollectionView.cs" />
-    <Compile Include="System.ComponentModel\ListSortDirection.cs" />
     <Compile Include="System.ComponentModel\PageChangingEventArgs.cs" />
     <Compile Include="System.ComponentModel\ReadOnlyAttribute.cs" />
-    <Compile Include="System.ComponentModel\WORKINPROGRESS\CurrentChangingEventArgs.cs" />
-    <Compile Include="System.ComponentModel\WORKINPROGRESS\CurrentChangingEventHandler.cs" />
     <Compile Include="System.ComponentModel\WORKINPROGRESS\GroupDescription.cs" />
-    <Compile Include="System.ComponentModel\WORKINPROGRESS\ICollectionView.cs" />
     <Compile Include="System.ComponentModel\WORKINPROGRESS\IEditableCollectionView.cs" />
-    <Compile Include="System.ComponentModel\WORKINPROGRESS\ISupportInitialize.cs" />
     <Compile Include="System.ComponentModel\WORKINPROGRESS\NewItemPlaceholderPosition.cs" />
-    <Compile Include="System.ComponentModel\WORKINPROGRESS\SortDescription.cs" />
-    <Compile Include="System.ComponentModel\WORKINPROGRESS\SortDescriptionCollection.cs" />
-    <Compile Include="System.IO.IsolatedStorage\WORKINPROGRESS\IsolatedStorageSettings.cs" />
     <Compile Include="System.Net\WORKINPROGRESS\HttpWebRequest.cs" />
     <Compile Include="System.Net\WORKINPROGRESS\WebRequest.cs" />
     <Compile Include="System.Resources\ResourceManager.cs" />
@@ -83,12 +74,8 @@
     <Compile Include="System.Windows.Interactivity\WORKINPROGRESS\TriggerAction.cs" />
     <Compile Include="System.Windows.Interactivity\WORKINPROGRESS\TriggerAction_1.cs" />
     <Compile Include="System.Windows.Markup\WORKINPROGRESS\XmlLanguage.cs" />
-    <Compile Include="Windows.UI.Xaml.Controls\WORKINPROGRESS\DescriptionViewer.cs" />
     <Compile Include="Windows.UI.Xaml.Controls\WORKINPROGRESS\Frame.cs" />
-    <Compile Include="Windows.UI.Xaml.Controls\WORKINPROGRESS\LayoutTransformer.cs" />
-    <Compile Include="Windows.UI.Xaml.Controls\WORKINPROGRESS\Page.cs" />
     <Compile Include="Windows.UI.Xaml.Data\WORKINPROGRESS\BindingExpressionBase.cs" />
-    <Compile Include="Windows.UI.Xaml.Data\WORKINPROGRESS\CollectionViewSource.cs" />
     <Compile Include="Windows.UI.Xaml.Input\WORKINPROGRESS\TextCompositionEventHandler.cs" />
     <Compile Include="Windows.UI.Xaml.Navigation\WORKINPROGRESS\LoadCompletedEventHandler.cs" />
     <Compile Include="System.Xml.Linq\XObject.cs" />
@@ -745,7 +732,6 @@
     <Compile Include="Windows.UI.Xaml.Navigation\WORKINPROGRESS\NavigatingCancelEventArgs.cs" />
     <Compile Include="Windows.UI.Xaml.Navigation\WORKINPROGRESS\NavigatingCancelEventHandler.cs" />
     <Compile Include="Windows.UI.Xaml.Navigation\WORKINPROGRESS\NavigationMode.cs" />
-    <Compile Include="Windows.UI.Xaml.Navigation\WORKINPROGRESS\NavigationService.cs" />
     <Compile Include="Windows.UI.Xaml.Navigation\WORKINPROGRESS\NavigationStoppedEventHandler.cs" />
     <Compile Include="Windows.UI.Xaml.Shapes\Ellipse.cs" />
     <Compile Include="Windows.UI.Xaml.Shapes\Line.cs" />
@@ -827,7 +813,6 @@
     <Compile Include="Windows.UI.Xaml\WhenToCallPropertyChangedEnum.cs" />
     <Compile Include="Windows.UI.Xaml\Window.cs" />
     <Compile Include="Windows.UI.Xaml\WindowSizeChangedEventHandler.cs" />
-    <Compile Include="Windows.UI.Xaml\WORKINPROGRESS\Application.cs" />
     <Compile Include="Windows.UI.Xaml\WORKINPROGRESS\ApplicationUnhandledExceptionEventArgs.cs" />
     <Compile Include="Windows.UI.Xaml\WORKINPROGRESS\AssemblyPart.cs" />
     <Compile Include="Windows.UI.Xaml\WORKINPROGRESS\AssemblyPartCollection.cs" />
