--- src/CSHTML5.Runtime/System.ComponentModel/ListSortDirection.cs
@@ -1,17 +1,0 @@
-﻿namespace System.ComponentModel
-{
-    //
-    // Summary:
-    //     Specifies the direction of a sort operation.
-    public enum ListSortDirection
-    {
-        //
-        // Summary:
-        //     Sorts in ascending order.
-        Ascending = 0,
-        //
-        // Summary:
-        //     Sorts in descending order.
-        Descending = 1
-    }
-}
