--- src/CSHTML5.Runtime/System.ComponentModel/WORKINPROGRESS/CurrentChangingEventHandler.cs
@@ -1,19 +1,0 @@
-﻿#if WORKINPROGRESS
-
-namespace System.ComponentModel
-{
-    //
-    // Summary:
-    //     Represents a method that can handle the System.ComponentModel.ICollectionView.CurrentChanging
-    //     event of an System.ComponentModel.ICollectionView implementation.
-    //
-    // Parameters:
-    //   sender:
-    //     The source of the event.
-    //
-    //   e:
-    //     The event data.
-    public delegate void CurrentChangingEventHandler(object sender, CurrentChangingEventArgs e);
-}
-
-#endif
