--- src/CSHTML5.Runtime/System.IO.IsolatedStorage/WORKINPROGRESS/IsolatedStorageSettings.cs
@@ -1,14 +1,0 @@
-﻿#if WORKINPROGRESS
-
-namespace System.IO.IsolatedStorage
-{
-    public sealed partial class IsolatedStorageSettings
-    {
-        public void Save()
-        {
-
-        }
-    }
-}
-
-#endif
