+++ src/CSHTML5.Runtime/System.ServiceModel/ServiceKnownTypeAttribute.cs
@@ -16,7 +16,7 @@
 //===============================================================================
 
 
-#if !OPENSILVER
+
 #if WCF_STACK || BRIDGE || CSHTML5BLAZOR
 
 namespace System.ServiceModel
@@ -66,5 +66,4 @@ namespace System.ServiceModel
     }
 }
 
-#endif
 #endif
