+++ src/CSHTML5.Runtime/System.Windows/Clipboard.cs
@@ -35,6 +35,7 @@ namespace System.Windows
         public static void SetText(string text)
         {
             // Credits: https://stackoverflow.com/questions/400212/how-do-i-copy-to-the-clipboard-in-javascript
+
 #if !CSHTML5NETSTANDARD
             if (CSHTML5.Interop.IsRunningInTheSimulator)
             {
@@ -43,7 +44,7 @@ namespace System.Windows
             else
             {
 #endif
-            CSHTML5.Interop.ExecuteJavaScript(@"
+                CSHTML5.Interop.ExecuteJavaScript(@"
                   // IE specific
                   if (window.clipboardData && window.clipboardData.setData) {
                     return clipboardData.setData(""Text"", $0);
