+++ src/CSHTML5.Runtime/Windows.UI.Xaml.Controls/WORKINPROGRESS/Frame.cs
@@ -3,7 +3,6 @@
 #if MIGRATION
 using System.Windows.Navigation;
 #else
-using System;
 using Windows.UI.Xaml.Navigation;
 #endif
