+++ src/CSHTML5.Runtime/Windows.UI.Xaml.Media.Animation/WORKINPROGRESS/BackEase.cs
@@ -1,6 +1,4 @@
-﻿#if WORKINPROGRESS
-
-using System;
+﻿using System;
 
 #if MIGRATION
 namespace System.Windows.Media.Animation
@@ -8,6 +6,10 @@ namespace System.Windows.Media.Animation
 namespace Windows.UI.Xaml.Media.Animation
 #endif
 {
+
+#if WORKINPROGRESS
+    #region Not Supported yet
+
     /// <summary>
     /// This class implements an easing function that backs up before going to the destination.
     /// </summary>
@@ -15,6 +17,7 @@ namespace Windows.UI.Xaml.Media.Animation
     {
         public double Amplitude { get; set; }
     }
-}
 
+    #endregion
 #endif
+}
