+++ src/CSHTML5.Runtime/Windows.UI.Xaml.Navigation/WORKINPROGRESS/NavigatingCancelEventHandler.cs
@@ -19,7 +19,7 @@ namespace Windows.UI.Xaml.Navigation
     //
     //   e:
     //     The data for the event.
-    //[TypeForwardedFrom("System.Windows.Controls.Navigation, Version=2.0.5.0, Culture=Neutral, PublicKeyToken=31bf3856ad364e35")]
+    [TypeForwardedFrom("System.Windows.Controls.Navigation, Version=2.0.5.0, Culture=Neutral, PublicKeyToken=31bf3856ad364e35")]
     public delegate void NavigatingCancelEventHandler(object sender, NavigatingCancelEventArgs e);
 }
