--- src/CSHTML5.Runtime/Windows.UI.Xaml.Navigation/WORKINPROGRESS/NavigationService.cs
@@ -1,20 +1,0 @@
-﻿#if WORKINPROGRESS
-
-#if MIGRATION
-namespace System.Windows.Navigation
-#else
-namespace Windows.UI.Xaml.Navigation
-#endif
-{
-    public sealed class NavigationService
-    {
-        public bool CanGoBack { get; }
-
-        public void GoBack()
-        {
-
-        }
-    }
-}
-
-#endif
