+++ src/CSHTML5.Runtime/Windows.UI.Xaml/MessageBoxResult.cs
@@ -42,17 +42,5 @@ namespace Windows.UI.Xaml
         /// The user clicked the Cancel button or pressed ESC.
         /// </summary>
         Cancel = 2,
-#if WORKINPROGRESS
-        /// <summary>
-        /// The dialog box return value is
-        /// Yes (usually sent from a button labeled Yes).
-        /// </summary>
-        Yes,
-        /// <summary>
-        /// The dialog box return value is
-        /// No (usually sent from a button labeled No).
-        /// </summary>
-        No,
-#endif
     }
 }
